Move 'name/italic_names' to Universal profile.
This check uses the filename to determine whether the font should have various italic bits set.
File naming is a matter for the OpenType specification, so this should be under Universal.

(https://github.com/fonttools/fontbakery/issues/4858)
