+++ profile-universal/src/checks/mod.rs
@@ -1,5 +1,6 @@
 pub mod arabic_spacing_symbols;
 pub mod glyphnames;
+pub mod name_italic_names;
 pub mod name_no_copyright_on_description;
 pub mod name_trailing_spaces;
 pub mod required_tables;
